SPTH: Created guide item for new player inventory
The item is an Artifact called "Caged demon called Logon the Guide".
When the player Applies the item, it calls the code in the method called use_speechTherapyGame_Guide().

When the player Invokes the item, it tames adjacent monsters.

When the player Applies the item, it speaks a random Hint. This should be changed so that it speaks speech-therapy related hints.