SPTH: Added interprocess communication (pipes)
This commit adds basic interprocess communication using Windows Named Pipes. This is only designed to work on Windows.

It uses the pipe name: articulate_assistant_pipe_game

On Applying the item Logon the Guide, it sends a string to the Named Pipe and waits for a one-byte response which it displays on-screen. This is designed to pair with the named pipe served by AAA.